feat: BiFMIndex carries the number of steps applied in the Index
with this conversion from begin/end of a query can be done.
This is especially interesting if the begin is reported, but one is
interested in the ending. This is for example the case when storing the
reverse (complement) and it requires to correct the position in the text
which might report the 'ending' since its saved reversed.